# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  model.Brand:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Buyer:
    properties:
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
    type: object
...
  model.Category:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.CurrentItemInfo:
    properties:
      article:
        type: string
      brand:
        type: string
      category:
        type: string
      description:
        type: string
      height:
        type: integer
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      length:
        type: integer
      material:
        type: string
      name:
        type: string
      price:
        type: number
      price_with_discount:
        type: number
      quantity:
        type: integer
      seller:
        type: string
      seller_id:
        type: string
      weight:
        type: integer
      width:
        type: integer
    type: object
  model.FilterRequest:
    properties:
      brands:
        items:
          type: integer
        type: array
      categories:
        items:
          type: integer
        type: array
      materials:
        items:
          type: integer
        type: array
      max_price:
        type: number
      min_price:
        type: number
      sellers:
        items:
          type: integer
        type: array
    type: object
  model.Image:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      url:
        type: string
    type: object
  model.Item:
    properties:
      article:
        type: string
      brand:
        $ref: '#/definitions/model.Brand'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      description:
        type: string
      favoritedBy:
        items:
          $ref: '#/definitions/model.Buyer'
        type: array
      height:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      length:
        type: integer
      material:
        $ref: '#/definitions/model.Material'
      material_id:
        type: integer
      name:
        type: string
      price:
        type: number
      price_with_discount:
        type: number
      quantity:
        type: integer
      seller:
        $ref: '#/definitions/model.Seller'
      seller_id:
        type: string
      weight:
        type: integer
      width:
        type: integer
    type: object
  model.ItemInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  model.Material:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
  model.MaterialInput:
    properties:
      name:
        type: string
    type: object
  model.MaterialOutput:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.Order:
    properties:
      buyer:
        $ref: '#/definitions/model.Buyer'
      buyer_id:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  model.OrderItem:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      order:
        $ref: '#/definitions/model.Order'
      order_id:
        type: integer
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
  model.Seller:
    properties:
      balance:
        type: number
      email:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
      password:
        type: string
      shop_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/material:
    delete:
      description: Delete a material from the system by its ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Material ID
        in: query
        name: material_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material deleted successfully
          schema:
            type: string
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete material
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a material by ID
      tags:
      - Materials
    post:
      consumes:
      - application/json
      description: Create a new material in the system
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Material data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.MaterialInput'
      produces:
      - application/json
      responses:
        "201":
          description: Material created successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create material
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new material
      tags:
      - Materials
...
  /item:
    get:
      description: Retrieve an item by its ID
      parameters:
      - description: Item ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item details
          schema:
            $ref: '#/definitions/model.CurrentItemInfo'
        "400":
          description: Invalid item ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get item by ID
      tags:
      - Items
    post:
      consumes:
      - application/json
      description: Retrieve a list of items, optionally filtered
      parameters:
      - description: Filter criteria
        in: body
        name: filters
        required: true
        schema:
          $ref: '#/definitions/model.FilterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of items
          schema:
            items:
              $ref: '#/definitions/model.ItemInfo'
            type: array
        "500":
          description: Failed to get items
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get item list
      tags:
      - Items
  /material:
    get:
      description: Retrieve a list of materials from the system
      produces:
      - application/json
      responses:
        "200":
          description: List of materials
          schema:
            items:
              $ref: '#/definitions/model.MaterialOutput'
            type: array
        "500":
          description: Failed to retrieve material list
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get material list
      tags:
      - Materials
...
  /seller/item:
    patch:
      consumes:
      - application/json
      description: Update an item by its ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Item ID
        in: query
        name: id
        required: true
        type: string
      - description: Item data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Item'
      produces:
      - application/json
      responses:
        "200":
          description: Item updated successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update item
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update an existing item
      tags:
      - Items
    post:
      consumes:
      - application/json
      description: Create a new item in the system
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Item data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Item created successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create item
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new item
      tags:
      - Items
  /seller/item/image:
    post:
      description: Upload an image for a specific item by ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Item ID
        in: query
        name: item_id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'url": "url'
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Invalid item ID or failed to upload image
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to save image
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Upload an image for an item
      tags:
      - Items
...
swagger: "2.0"